Refactor to require import
