+++ src/library/scala/util/package.scala
@@ -1,0 +1,8 @@
+package scala
+
+package object util {
+  /**
+   * Adds chaining methods `tap` and `pipe` to every type. See [[ChainingOps]].
+   */
+  object chainingOps extends ChainingSyntax
+}
