+++ test/junit/scala/PredefTest.scala
@@ -20,20 +20,4 @@ class PredefTest {
     assertEquals("bazzz", c)
     assertEquals(5, res)
   }
-
-  @Test
-  def testAnyTap: Unit = {
-    var x: Int = 0
-    val result = List(1, 2, 3)
-      .tap(xs => x = xs(0))
-    assertEquals(1, x)
-    assertEquals(List(1, 2, 3), result)
-  }
-
-  @Test
-  def testAnyPipe: Unit = {
-    val result = List(1, 2, 3)
-      .pipe(xs => xs.mkString(","))
-    assertEquals("1,2,3", result)
-  }
 }
